+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachangelog.com/)
 
 ## [Unreleased]
+### Added
+- check-consul-service-health and check-service-consul now accept a --tags filter @wg-tsuereth
 ### Fixed
 - check-service-consul should look for 'critical' instead of 'failing'